etl/requirements.txt
elasticsearch==8.9.0
psycopg2-binary==2.9.7
redis==5.0.0

backoff==2.2.1

pydantic==2.3.0
pydantic-settings==2.0.3

sentry-sdk==1.42.0
